fixup: rename fake2 to fake
