+++ crates/rpc/src/v02/method/get_state_update.rs
@@ -351,7 +351,7 @@ mod tests {
         // let mut connection = storage.connection().unwrap();
         // let tx = connection.transaction().unwrap();
         // let state_updates = pathfinder_storage::test_fixtures::init::with_n_state_updates(&tx, 3);
-        let state_updates = pathfinder_storage::fake2::with_n_blocks(&storage, 3)
+        let state_updates = pathfinder_storage::fake::with_n_blocks(&storage, 3)
             .into_iter()
             .map(|(_, _, x)| x.into())
             .collect();
