+++ crates/rpc/src/v03/method/get_state_update.rs
@@ -409,7 +409,7 @@ mod tests {
         // let state_updates = pathfinder_storage::test_fixtures::init::with_n_state_updates(&tx, 3);
         // tx.commit().unwrap();
 
-        let state_updates = pathfinder_storage::fake2::with_n_blocks(&storage, 3)
+        let state_updates = pathfinder_storage::fake::with_n_blocks(&storage, 3)
             .into_iter()
             .map(|(_, _, x)| x.into())
             .collect();
