+++ crates/storage/src/lib.rs
@@ -7,11 +7,9 @@ mod prelude;
 
 mod connection;
 #[cfg(any(feature = "test-utils", test))]
-pub mod fake2;
+pub mod fake;
 mod params;
 mod schema;
-// #[cfg(any(feature = "test-utils", test))]
-// pub mod test_fixtures;
 #[cfg(any(feature = "test-utils", test))]
 pub mod test_utils;
 
@@ -19,6 +17,7 @@ use std::num::NonZeroU32;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
+#[cfg(any(feature = "test-utils", test))]
 pub use connection::*;
 
 use pathfinder_common::{BlockHash, BlockNumber};
